feat: 결제 api 개발 중, Money 클래스 도입

+++ src/main/java/wayc/backend/shop/application/dto/response/find/FindItemsResponseDto.java
@@ -34,7 +34,7 @@ public class FindItemsResponseDto {
                 item.getId(),
                 item.getName(),
                 item.getShop().getName(),
-                item.getPrice(),
+                item.getPrice().intValue(),
                 item.getImageUrl(),
                 item.getCategory()
         );
